Refactored index.ts and added transformer unit test.
