+++ magda-ckan-connector/package.json
@@ -10,6 +10,7 @@
     "watch": "tsc -p tsconfig-build.json --watch",
     "start": "node dist/index.js",
     "dev": "run-typescript-in-nodemon src/index.ts",
+    "test": "mocha --compilers ts:ts-node/register,tsx:ts-node/register --userId dummy --jwtSecret dummy --sourceUrl dummy --id test --name test-name --require tsconfig-paths/register \"src/test/**/*.spec.ts\"",
     "docker-build-local": "create-docker-context-for-node-component --build --push --tag auto --local",
     "docker-build-prod": "create-docker-context-for-node-component --build --push --tag auto",
     "retag-and-push": "retag-and-push"
